+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This release marked the addition of strict typing and return type declarations (
 - Fixed specifying cell widths, background color, etc on `PhpOffice\PhpWord\Style\Cell` @0b10011 #1669
 - Escape arrays of replacements in `TemplateProcessor` @0b10011 #1669
 - Escape text provided for `<title>` when exporting to HTML @0b10011
+- Export `<h1>` instead of `<p class="Heading1">` for headings @0b10011 #1692
 
 ### Miscellaneous
 -